+++ src/test/java/com/github/toolarium/enumeration/configuration/processor/EnumConfigurationProcessorTest.java
@@ -10,14 +10,19 @@ import static com.google.testing.compile.CompilationSubject.assertThat;
 import static com.google.testing.compile.Compiler.javac;
 import static com.google.testing.compile.JavaFileObjects.forSourceString;
 import static com.google.testing.compile.JavaSourcesSubjectFactory.javaSources;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
 
+import com.github.toolarium.enumeration.configuration.dto.EnumConfigurations;
+import com.github.toolarium.enumeration.configuration.resource.EnumConfigurationResourceFactory;
 import com.github.toolarium.enumeration.configuration.util.JavaFileObjectUtil;
 import com.google.common.collect.ImmutableList;
 import com.google.testing.compile.Compilation;
 import com.google.testing.compile.Compiler;
 import com.google.testing.compile.JavaFileObjects;
+import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import javax.tools.JavaFileObject;
 import org.junit.jupiter.api.Test;
@@ -132,6 +137,50 @@ public class EnumConfigurationProcessorTest {
 
         //assertThat(compilation).hadErrorContaining("No types named HelloWorld!").inFile(helloWorld).onLine(23).atColumn(5);
         //assertThat(compilation).generatedSourceFile("GeneratedHelloWorld").hasSourceEquivalentTo(JavaFileObjects.forResource("GeneratedHelloWorld.java"));
+    }
+
+    
+    /**
+     * Annotation test with source reference
+     * 
+     * @throws IOException In case source could not be loaded 
+     */
+    @Test
+    public void compileAnnotationTestWithTag() throws IOException {
+        
+        /*
+        assertAbout(javaSources())
+        .that(ImmutableList.of(JavaFileObjectUtil.getInstance().loadTestSourceByClass(IMyEnumConfiguration.class),
+                               JavaFileObjectUtil.getInstance().loadTestSourceByClass(MyEnumWithTagConfiguration.class)))
+        .processedWith(new EnumConfigurationProcessor()).compilesWithoutError();
+        */
+
+        TestEnumConfigurationProcessor testEnumConfigurationProcessor = new TestEnumConfigurationProcessor();
+        assertAbout(javaSources())
+        .that(ImmutableList.of(JavaFileObjectUtil.getInstance().loadTestSourceByClass(IMyEnumConfiguration.class),
+                               JavaFileObjectUtil.getInstance().loadTestSourceByClass(MyEnumWithTagConfiguration.class)))
+        .processedWith(testEnumConfigurationProcessor).compilesWithoutError();
+
+        String content = testEnumConfigurationProcessor.getGeneratedEnumConfiguration();
+        assertNotNull(content);
+        EnumConfigurations enumConfigurations = EnumConfigurationResourceFactory.getInstance().load(new ByteArrayInputStream(content.getBytes()));
+        assertNotNull(enumConfigurations);
+        
+        assertEquals("myTag", enumConfigurations.getEnumConfigurationList().iterator().next().getTag());
+        
+        
+        testEnumConfigurationProcessor = new TestEnumConfigurationProcessor();
+        assertAbout(javaSources())
+        .that(ImmutableList.of(JavaFileObjectUtil.getInstance().loadTestSourceByClass(IMyEnumConfiguration.class),
+                               JavaFileObjectUtil.getInstance().loadTestSourceByClass(MyEnumConfiguration.class)))
+        .processedWith(testEnumConfigurationProcessor).compilesWithoutError();
+
+        content = testEnumConfigurationProcessor.getGeneratedEnumConfiguration();
+        assertNotNull(content);
+        enumConfigurations = EnumConfigurationResourceFactory.getInstance().load(new ByteArrayInputStream(content.getBytes()));
+        assertNotNull(enumConfigurations);
+        
+        assertNull(enumConfigurations.getEnumConfigurationList().iterator().next().getTag());
     }